unix socket path config
