+++ Package.swift
@@ -11,7 +11,7 @@ let package = Package(
     ],
     dependencies: [
         .package(url: "https://github.com/vapor/fluent-kit.git", from: "1.0.0-rc.1"),
-        .package(url: "https://github.com/vapor/mysql-kit.git", from: "4.0.0-rc.1"),
+        .package(url: "https://github.com/vapor/mysql-kit.git", from: "4.0.0-rc.1.2"),
         .package(url: "https://github.com/apple/swift-log.git", from: "1.0.0"),
     ],
     targets: [
