+++ Sources/FluentMySQLDriver/FluentMySQLError.swift
@@ -1,0 +1,3 @@
+enum FluentMySQLError: Error {
+    case invalidURL(String)
+}
